OpenAI/requirements.txt
openai==1.13.3
tiktoken==0.6.0
python-dotenv==1.0.1